Create README.md
I am new to the IT world, i am coming from a procurement background. It all started when i realize that the corona virus is not ready to go anytime soon and how much it affected the economy of the world and then i develop the interest of working from home or anywhere I am. At first i was going for data analysis so i started with learning python codes and then I realize I would do a lot better as a developer than a data analyst, so here I am working towards becoming a good wed developer.